expand fields available to linter
this expands the parsing logic to enable linting on additional
fields in the dashboard schema:

- .__inputs
- .panels[].targets[].datasource

additionally, the datasource now will have a 'type' property available,
if present in the dashboard model.

one target use-case for this is ensuring that datasources defined
in panels and targets are referencing datasources defined in the
__inputs section. the __inputs section is particularly important for
dashboards exported for sharing w/ other grafana environments.